requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
bcrypt==3.2.0
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.4
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.1.5
django-heroku==0.3.1
firebase-admin==4.3.0
google-api-core==1.22.1
google-api-python-client==1.10.0
google-auth==1.20.1
google-auth-httplib2==0.0.4
google-cloud-core==1.4.1
google-cloud-firestore==1.8.1
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-django==2.2.0
graphql-core==2.3.2
graphql-relay==2.0.1
grpcio==1.31.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
msgpack==1.0.0
mysqlclient==2.0.1
promise==2.3
protobuf==3.13.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
rsa==4.6
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
